Moved the database to another folder.
