+++ app/src/main/java/com/example/android/cryptoapp/data/local/db/CurrencyDao.kt
@@ -1,6 +1,5 @@
-package com.example.android.cryptoapp.database
+package com.example.android.cryptoapp.data.local.db
 
-import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
