+++ app/src/main/java/com/example/android/cryptoapp/data/local/db/CurrencyRoomDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.android.cryptoapp.database
+package com.example.android.cryptoapp.data.local.db
 
 import android.content.Context
 import androidx.room.Database
